Moving _deprecate_pos_args to internals and other cleanup
