+++ python/cuml/_thirdparty/sklearn/preprocessing/_data.py
@@ -43,12 +43,13 @@ from ....thirdparty_adapters.sparsefuncs_fast import \
     (inplace_csr_row_normalize_l1, inplace_csr_row_normalize_l2,
      csr_polynomial_expansion)
 from ....common.import_utils import check_cupy8
-from ....common.input_utils import _deprecate_pos_args
 
 from ....common.array import CumlArray
 from ....common.array_sparse import SparseCumlArray
 from ....common.array_descriptor import CumlArrayDescriptor
 from ....internals import api_return_generic
+from ....internals import _deprecate_pos_args
+
 from ....common.memory_utils import using_output_type
 
 BOUNDS_THRESHOLD = 1e-7
