+++ python/cuml/_thirdparty/sklearn/preprocessing/_imputation.py
@@ -26,9 +26,9 @@ from cuml.common.mixins import AllowNaNTagMixin, SparseInputTagMixin, \
 from ..utils.validation import check_is_fitted
 from ..utils.validation import FLOAT_DTYPES
 from ....common.import_utils import check_cupy8
-from ....common.input_utils import _deprecate_pos_args
 from ....common.array_sparse import SparseCumlArray
 from ....common.array_descriptor import CumlArrayDescriptor
+from ....internals import _deprecate_pos_args
 
 
 def is_scalar_nan(x):
