+++ python/cuml/common/input_utils.py
@@ -16,9 +16,6 @@
 
 import copy
 from collections import namedtuple
-from inspect import Parameter, signature
-import typing
-import warnings
 
 import cudf
 import cupy as cp
@@ -26,7 +23,6 @@ import cupyx
 import numba.cuda
 import numpy as np
 import pandas as pd
-from cuml.common.type_utils import _DecoratorType, wraps_typed
 import cuml.internals
 import cuml.common.array
 from cuml.common.array import CumlArray
@@ -613,70 +609,3 @@ def sparse_scipy_to_cp(sp, dtype):
     v = cp.asarray(values, dtype=dtype)
 
     return cupyx.scipy.sparse.coo_matrix((v, (r, c)), sp.shape)
-
-
-class _deprecate_pos_args:
-    """
-    Decorator that issues a warning when using positional args that should be
-    keyword args. Mimics sklearn's `_deprecate_positional_args` with added
-    functionality.
-
-    For any class that derives from `cuml.Base`, this decorator will be
-    automatically added to `__init__`. In this scenario, its assumed that all
-    arguments are keyword arguments. To override the functionality this
-    decorator can be manually added, allowing positional arguments if
-    necessary.
-
-    Parameters
-    ----------
-    version : str
-        This version will be specified in the warning message as the
-        version when positional arguments will be removed
-
-    """
-
-    FLAG_NAME: typing.ClassVar[str] = "__cuml_deprecated_pos"
-
-    def __init__(self, version: str):
-
-        self._version = version
-
-    def __call__(self, func: _DecoratorType) -> _DecoratorType:
-
-        sig = signature(func)
-        kwonly_args = []
-        all_args = []
-
-        # Store all the positional and keyword only args
-        for name, param in sig.parameters.items():
-            if param.kind == Parameter.POSITIONAL_OR_KEYWORD:
-                all_args.append(name)
-            elif param.kind == Parameter.KEYWORD_ONLY:
-                kwonly_args.append(name)
-
-        @wraps_typed(func)
-        def inner_f(*args, **kwargs):
-            extra_args = len(args) - len(all_args)
-            if extra_args > 0:
-                # ignore first 'self' argument for instance methods
-                args_msg = [
-                    '{}={}'.format(name, arg) for name,
-                    arg in zip(kwonly_args[:extra_args], args[-extra_args:])
-                ]
-                warnings.warn(
-                    "Pass {} as keyword args. From version {}, "
-                    "passing these as positional arguments will "
-                    "result in an error".format(", ".join(args_msg),
-                                                self._version),
-                    FutureWarning,
-                    stacklevel=2)
-
-            # Convert all positional args to keyword
-            kwargs.update({k: arg for k, arg in zip(sig.parameters, args)})
-
-            return func(**kwargs)
-
-        # Set this flag to prevent auto adding this decorator twice
-        inner_f.__dict__[_deprecate_pos_args.FLAG_NAME] = True
-
-        return inner_f
