+++ python/cuml/tsa/auto_arima.pyx
@@ -32,7 +32,7 @@ from cuml.common import logger
 from cuml.common.array_descriptor import CumlArrayDescriptor
 from cuml.common.array import CumlArray
 from cuml.common.base import Base
-from cuml.common.input_utils import _deprecate_pos_args
+from cuml.internals import _deprecate_pos_args
 from cuml.raft.common.handle cimport handle_t
 from cuml.raft.common.handle import Handle
 from cuml.common import input_to_cuml_array
